Refactored api file to be inside of config foler
