+++ Android/indoor_localization/lib/data/datasources/remote/asset_remote_datasource.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'package:indoor_localization/data/connections/api.dart';
+import 'package:indoor_localization/config/api.dart';
 
 class AssetRemoteDataSource {
   final http.Client client;
